GLM removed

+++ sources/etx/render/host/scene_loader.hxx
@@ -27,8 +27,8 @@ struct SceneRepresentation {
   ETX_DECLARE_PIMPL(SceneRepresentation, 2048);
 };
 
-Camera build_camera(const float3& origin, const float3& target, const float3& up, const float2& viewport, float fov, float lens_radius, float focal_distance);
-void update_camera(Camera& camera, const float3& origin, const float3& target, const float3& up, const float2& viewport, float fov);
+Camera build_camera(const float3& origin, const float3& target, const float3& up, const uint2& viewport, float fov, float lens_radius, float focal_distance);
+void update_camera(Camera& camera, const float3& origin, const float3& target, const float3& up, const uint2& viewport, float fov);
 float get_camera_fov(const Camera& camera);
 
 }  // namespace etx
